+++ src/multipliers.lua
@@ -16,9 +16,10 @@ game.TraitData.ManaBurstBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.FocusRawDamageBoon.RarityLevels.Perfect = { Multiplier = 15 / 5 } -- 2/5
 
 -- apollo
-game.TraitData.ApolloWeaponBoon.RarityLevels.Perfect = { Multiplier = 2.25 }
+game.TraitData.ApolloWeaponBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.ApolloSpecialBoon.RarityLevels.Perfect = { Multiplier = 2.67 }
 game.TraitData.ApolloCastBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
+game.TraitData.ApolloExCastBoon.RarityLevels.Perfect = { Multiplier = 2.0 }
 game.TraitData.ApolloSprintBoon.RarityLevels.Perfect = { Multiplier = 8 / 3 } -- 1/3
 game.TraitData.ApolloManaBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.ApolloRetaliateBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
@@ -34,7 +35,7 @@ game.TraitData.DemeterWeaponBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.DemeterSpecialBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.DemeterCastBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.DemeterSprintBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
-game.TraitData.DemeterManaBoon.RarityLevels.Perfect = { Multiplier = 75 / 25 } -- 10/25
+game.TraitData.DemeterManaBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.CastNovaBoon.RarityLevels.Perfect = { Multiplier = 6 }          -- 1
 game.TraitData.PlantHealthBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.BoonGrowthBoon.RarityLevels.Perfect = { Multiplier = 6 }
@@ -45,54 +46,60 @@ game.TraitData.RootDurationBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 
 -- hephaestus
 game.TraitData.HephaestusWeaponBoon.RarityLevels.Perfect = { Multiplier = 2 / 12 }   -- -2/12
-game.TraitData.HephaestusSpecialBoon.RarityLevels.Perfect = { Multiplier = 10 / 20 } -- -2/20
+game.TraitData.HephaestusSpecialBoon.RarityLevels.Perfect = { Multiplier = 4.5/16 } -- -3/16, then -2/16. let's meet halfway
 game.TraitData.HephaestusCastBoon.RarityLevels.Perfect = { Multiplier = 2.2 }        -- 0.2
 game.TraitData.HephaestusSprintBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.HephaestusManaBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
-game.TraitData.ChargeCounterBoon.RarityLevels.Perfect = { Multiplier = 2.25 }
+--game.TraitData.ChargeCounterBoon.RarityLevels.Perfect = { Multiplier = 2.25 } boon removed 6/26
 game.TraitData.AntiArmorBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.HeavyArmorBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.ArmorBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.EncounterStartDefenseBuffBoon.RarityLevels.Perfect = { Multiplier = 2.25 }
 game.TraitData.ManaToHealthBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
+game.TraitData.MassiveDamageBoon.RarityLevels.Heroic = { Multiplier = 2.5 } -- appears to have incorrect value in 6/26 patch. i shall fix
+game.TraitData.MassiveDamageBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.MassiveKnockupBoon.RarityLevels.Perfect = { Multiplier = 8 / 3 } -- 1/3
 
 -- hera
 game.TraitData.HeraWeaponBoon.RarityLevels.Perfect = { Multiplier = 2 }
 game.TraitData.HeraSpecialBoon.RarityLevels.Perfect = { Multiplier = 1.1 / 0.6 } -- 0.1/0.6 (ok but why)
-game.TraitData.HeraCastBoon.RarityLevels.Perfect = { Multiplier = 2.925 }        -- 0.425
-game.TraitData.HeraSprintBoon.RarityLevels.Perfect = { Multiplier = 6 }
-game.TraitData.HeraManaBoon.RarityLevels.Perfect = { Multiplier = 0 }            -- -0.2
+game.TraitData.HeraCastBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
+game.TraitData.HeraSprintBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
+game.TraitData.HeraManaBoon.RarityLevels.Perfect = { Multiplier = 0.5 } -- -0.1
 game.TraitData.CommonGlobalDamageBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.DamageShareRetaliateBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.SwapBonusBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.BoonDecayBoon.RarityLevels.Perfect = { Multiplier = 2.2 }
+game.TraitData.OmegaHeraProjectileBoon.RarityLevels.Perfect = { Multiplier = 2.0 }
 game.TraitData.DamageSharePotencyBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.LinkedDeathDamageBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
-game.TraitData.FullManaExBoostBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
+--game.TraitData.FullManaExBoostBoon.RarityLevels.Perfect = { Multiplier = 3.5 } boon removed 6/26
 
 -- hestia
-game.TraitData.HestiaWeaponBoon.RarityLevels.Perfect = { Multiplier = 2.25 }
-game.TraitData.HestiaSpecialBoon.RarityLevels.Perfect = { Multiplier = 2.67 }
+game.TraitData.HestiaWeaponBoon.RarityLevels.Perfect = { Multiplier = 2.67 } -- 0.33ish
+game.TraitData.HestiaSpecialBoon.RarityLevels.Perfect = { Multiplier = 2.25 }
 game.TraitData.HestiaCastBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.HestiaSprintBoon.RarityLevels.Perfect = { Multiplier = 6 }
-game.TraitData.HestiaManaBoon.RarityLevels.Perfect = { Multiplier = 22 / 7 } -- 3/7
+game.TraitData.HestiaManaBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.SacrificeBoon.RarityLevels.Perfect = { Multiplier = 2.0 }
 game.TraitData.OmegaZeroBurnBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
-game.TraitData.CastProjectileBoon.RarityLevels.Perfect = { Multiplier = 3 } -- 0.4
+game.TraitData.CastProjectileBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.FireballManaSpecialBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.BurnExplodeBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
-game.TraitData.BurnConsumeBoon.RarityLevels.Perfect = { Multiplier = 2.25 }
+--game.TraitData.BurnConsumeBoon.RarityLevels.Perfect = { Multiplier = 2.25 } boon removed 6/26
 game.TraitData.BurnArmorBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
+game.TraitData.BurnStackBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 
 -- poseidon
 game.TraitData.PoseidonWeaponBoon.RarityLevels.Perfect = { Multiplier = 2.67 }
 game.TraitData.PoseidonSpecialBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.PoseidonCastBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
+game.TraitData.PoseidonExCastBoon.RarityLevels.Perfect = { Multiplier = 2.0 }
+game.TraitData.OmegaPoseidonProjectileBoon.RarityLevels.Perfect = { Multiplier = 2.0 }
 game.TraitData.PoseidonSprintBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.PoseidonManaBoon.RarityLevels.Perfect = { Multiplier = 2.0 }
 game.TraitData.EncounterStartOffenseBuffBoon.RarityLevels.Perfect = { Multiplier = 2.67 }
-game.TraitData.MinorLootBoon.RarityLevels.Perfect = { Multiplier = 2.5 } -- 0.3
+--game.TraitData.MinorLootBoon.RarityLevels.Perfect = { Multiplier = 2.5 } -- 0.3 boon removed 6/26
 game.TraitData.RoomRewardBonusBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.FocusDamageShaveBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.DoubleRewardBoon.RarityLevels.Perfect = { Multiplier = 2.25 }
@@ -101,12 +108,13 @@ game.TraitData.HeraWeaponBoon.RarityLevels.Perfect = { Multiplier = 6 }
 -- zeus
 game.TraitData.ZeusWeaponBoon.RarityLevels.Perfect = { Multiplier = 2.8 } -- 0.4
 game.TraitData.ZeusSpecialBoon.RarityLevels.Perfect = { Multiplier = 3 }
-game.TraitData.ZeusCastBoon.RarityLevels.Perfect = { Multiplier = 2.667 } -- 0.333
+game.TraitData.ZeusCastBoon.RarityLevels.Perfect = { Multiplier = 2.25 }
+game.TraitData.ZeusExCastBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.ZeusSprintBoon.RarityLevels.Perfect = { Multiplier = 2.25 }
-game.TraitData.ZeusManaBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
+game.TraitData.ZeusManaBoon.RarityLevels.Perfect = { Multiplier = 5/10 } -- -1/10
 game.TraitData.ZeusManaBoltBoon.RarityLevels.Perfect = { Multiplier = 2.25 }
 game.TraitData.CastAnywhereBoon.RarityLevels.Perfect = { Multiplier = 3 }
 game.TraitData.FocusLightningBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
-game.TraitData.DoubleBoltBoon.RarityLevels.Perfect = { Multiplier = 6 }
+game.TraitData.DoubleBoltBoon.RarityLevels.Perfect = { Multiplier = 3.5 }
 game.TraitData.EchoExpirationBoon.RarityLevels.Perfect = { Multiplier = 2 / 0.75 } -- 0.25/0.75
 game.TraitData.LightningDebuffGeneratorBoon.RarityLevels.Perfect = { Multiplier = 3.5 }